# doc/api/openapi.yaml
"openapi": "3.1.0"
"info":
  "title": "PixelPast"
  "version": "0.1.0"
"paths":
...
  "/api/days/context":
    "get":
      "tags":
        - "timeline"
      "summary": "Get day context preload"
      "description": "Return per-day context panels for an inclusive UTC range so the UI can preload hover data, tags, people, map points, and summary counts."
      "operationId": "get_day_context_api_days_context_get"
      "parameters":
        -
          "name": "start"
          "in": "query"
          "required": true
          "schema":
            "type": "string"
            "format": "date"
            "description": "Inclusive UTC start date for the requested context preload."
            "title": "Start"
          "description": "Inclusive UTC start date for the requested context preload."
          "examples":
            "trip_window_start":
              "summary": "Preload start date"
              "value": "2026-07-01"
        -
          "name": "end"
          "in": "query"
          "required": true
          "schema":
            "type": "string"
            "format": "date"
            "description": "Inclusive UTC end date for the requested context preload."
            "title": "End"
          "description": "Inclusive UTC end date for the requested context preload."
          "examples":
            "trip_window_end":
              "summary": "Preload end date"
              "value": "2026-07-03"
        -
          "name": "view_mode"
          "in": "query"
          "required": false
          "schema":
            "type": "string"
            "description": "Backend-defined daily view identifier that determines which derived day rows provide the context payload."
            "default": "activity"
            "title": "View Mode"
          "description": "Backend-defined daily view identifier that determines which derived day rows provide the context payload."
          "examples":
            "activity_view":
              "summary": "Default activity preload"
              "value": "activity"
        -
          "name": "person_ids"
          "in": "query"
          "required": false
          "schema":
            "type": "array"
            "items":
              "type": "integer"
            "description": "Repeatable person identifiers used to filter contributing canonical items."
            "default": []
            "title": "Person Ids"
          "description": "Repeatable person identifiers used to filter contributing canonical items."
          "examples":
            "anna":
              "summary": "Preload only days involving Anna Becker"
              "value": 7
        -
          "name": "tag_paths"
          "in": "query"
          "required": false
          "schema":
            "type": "array"
            "items":
              "type": "string"
            "description": "Repeatable normalized tag paths used to filter contributing canonical items."
            "default": []
            "title": "Tag Paths"
          "description": "Repeatable normalized tag paths used to filter contributing canonical items."
          "examples":
            "travel_tag":
              "summary": "Preload only travel-tagged days"
              "value": "travel/italy/venice"
        -
          "name": "location_geometry"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              -
                "type": "string"
              -
                "type": "null"
            "description": "Optional serialized geometry filter interpreted by the active projection provider."
            "title": "Location Geometry"
          "description": "Optional serialized geometry filter interpreted by the active projection provider."
          "examples":
            "venice_bbox":
              "summary": "Bounding box filter"
              "value": "bbox:12.285,45.423,12.380,45.465"
        -
          "name": "distance_latitude"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              -
                "type": "number"
                "maximum": 90
                "minimum": -90
              -
                "type": "null"
            "description": "Latitude for a radial distance filter. Use together with distance_longitude and distance_radius_meters."
            "title": "Distance Latitude"
          "description": "Latitude for a radial distance filter. Use together with distance_longitude and distance_radius_meters."
          "examples":
            "venice_center":
              "summary": "Radial filter latitude"
              "value": 45.4371
        -
          "name": "distance_longitude"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              -
                "type": "number"
                "maximum": 180
                "minimum": -180
              -
                "type": "null"
            "description": "Longitude for a radial distance filter. Use together with distance_latitude and distance_radius_meters."
            "title": "Distance Longitude"
          "description": "Longitude for a radial distance filter. Use together with distance_latitude and distance_radius_meters."
          "examples":
            "venice_center":
              "summary": "Radial filter longitude"
              "value": 12.3327
        -
          "name": "distance_radius_meters"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              -
                "type": "integer"
                "minimum": 1
              -
                "type": "null"
            "description": "Radius in meters for the radial distance filter."
            "title": "Distance Radius Meters"
          "description": "Radius in meters for the radial distance filter."
          "examples":
            "walkable_radius":
              "summary": "Five kilometer radial filter"
              "value": 5000
        -
          "name": "filename_query"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              -
                "type": "string"
                "minLength": 1
              -
                "type": "null"
            "description": "Case-insensitive filename search term applied to contributing assets."
            "title": "Filename Query"
          "description": "Case-insensitive filename search term applied to contributing assets."
          "examples":
            "camera_roll":
              "summary": "Filename substring match"
              "value": "IMG_20260703"
      "responses":
        "200":
          "description": "Dense day-context payload for the requested UTC range."
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/DayContextResponse"
              "examples":
                "preloaded_trip_week":
                  "summary": "Context preload for a short trip window"
                  "value":
                    "range":
                      "start": "2026-07-01"
                      "end": "2026-07-03"
                    "days":
                      -
                        "date": "2026-07-01"
                        "persons": []
                        "tags": []
                        "map_points": []
                        "summary_counts":
                          "events": 0
                          "assets": 0
                          "places": 0
                      -
                        "date": "2026-07-02"
                        "persons":
                          -
                            "id": 7
                            "name": "Anna Becker"
                            "role": "family"
                          -
                            "id": 12
                            "name": "Milo Tan"
                            "role": "friend"
                        "tags":
                          -
                            "path": "travel/italy/venice"
                            "label": "Venice"
                          -
                            "path": "activity/walking"
                            "label": "Walking"
                        "map_points":
                          -
                            "id": "location:2026-07-02:1"
                            "label": "Santa Lucia Station"
                            "latitude": 45.4419
                            "longitude": 12.3211
                        "summary_counts":
                          "events": 2
                          "assets": 1
                          "places": 1
        "400":
          "description": "The request was syntactically valid but violated a documented PixelPast API contract constraint."
          "content":
            "application/json":
              "examples":
                "invalid_range":
                  "summary": "Start date after end date"
                  "value":
                    "detail": "start must be less than or equal to end"
                "partial_range":
                  "summary": "Only one range boundary provided"
                  "value":
                    "detail": "start and end must both be provided together"
                "window_limit":
                  "summary": "Requested preload window too large"
                  "value":
                    "detail": "requested range exceeds maximum day context window of 31 days"
                "unsupported_view_mode":
                  "summary": "Unknown daily view identifier"
                  "value":
                    "detail": "unsupported view_mode: travel-intensity"
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
        "422":
          "description": "The request could not be parsed or failed declared parameter validation."
          "content":
            "application/json":
              "examples":
                "invalid_date":
                  "summary": "Unparseable ISO date"
                  "value":
                    "detail":
                      -
                        "type": "date_from_datetime_parsing"
                        "loc":
                          - "query"
                          - "start"
                        "msg": "Input should be a valid date or datetime, input is too short"
                        "input": "2026-02"
                "invalid_path_date":
                  "summary": "Invalid path day format"
                  "value":
                    "detail":
                      -
                        "type": "date_from_datetime_parsing"
                        "loc":
                          - "path"
                          - "day"
                        "msg": "Input should be a valid date or datetime, invalid date separator"
                        "input": "2026/07/14"
              "schema":
                "$ref": "#/components/schemas/ApiValidationErrorResponse"
...
"components":
  "schemas":
    "ApiErrorResponse":
      "properties":
        "detail":
          "type": "string"
          "title": "Detail"
      "type": "object"
      "required":
        - "detail"
      "title": "ApiErrorResponse"
      "description": "Error payload for contract-level request rejections."
    "ApiValidationErrorItem":
      "properties":
        "type":
          "type": "string"
          "title": "Type"
        "loc":
          "items":
            "anyOf":
              -
                "type": "string"
              -
                "type": "integer"
          "type": "array"
          "title": "Loc"
        "msg":
          "type": "string"
          "title": "Msg"
        "input":
          "anyOf":
            - {}
            -
              "type": "null"
          "title": "Input"
      "type": "object"
      "required":
        - "type"
        - "loc"
        - "msg"
      "title": "ApiValidationErrorItem"
      "description": "One FastAPI request validation error entry."
    "ApiValidationErrorResponse":
      "properties":
        "detail":
          "items":
            "$ref": "#/components/schemas/ApiValidationErrorItem"
          "type": "array"
          "title": "Detail"
      "type": "object"
      "required":
        - "detail"
      "title": "ApiValidationErrorResponse"
      "description": "Error payload returned when declared request validation fails."
...
    "DayContextDay":
      "properties":
        "date":
          "type": "string"
          "format": "date"
          "title": "Date"
        "persons":
          "items":
            "$ref": "#/components/schemas/ExplorationPerson"
          "type": "array"
          "title": "Persons"
        "tags":
          "items":
            "$ref": "#/components/schemas/ExplorationTag"
          "type": "array"
          "title": "Tags"
        "map_points":
          "items":
            "$ref": "#/components/schemas/DayContextMapPoint"
          "type": "array"
          "title": "Map Points"
        "summary_counts":
          "$ref": "#/components/schemas/DayContextSummaryCounts"
      "type": "object"
      "required":
        - "date"
        - "persons"
        - "tags"
        - "map_points"
        - "summary_counts"
      "title": "DayContextDay"
      "description": "Dense hover-context payload for one UTC calendar day."
    "DayContextMapPoint":
      "properties":
        "id":
          "type": "string"
          "title": "Id"
        "label":
          "type": "string"
          "title": "Label"
        "latitude":
          "type": "number"
          "title": "Latitude"
        "longitude":
          "type": "number"
          "title": "Longitude"
      "type": "object"
      "required":
        - "id"
        - "label"
        - "latitude"
        - "longitude"
      "title": "DayContextMapPoint"
      "description": "Serialized real-world map point for one UTC day."
    "DayContextResponse":
      "properties":
        "range":
          "$ref": "#/components/schemas/ExplorationRange"
        "days":
          "items":
            "$ref": "#/components/schemas/DayContextDay"
          "type": "array"
          "title": "Days"
      "type": "object"
      "required":
        - "range"
        - "days"
      "title": "DayContextResponse"
      "description": "Serialized response for an inclusive day-context preload range."
    "DayContextSummaryCounts":
      "properties":
        "events":
          "type": "integer"
          "title": "Events"
        "assets":
          "type": "integer"
          "title": "Assets"
        "places":
          "type": "integer"
          "title": "Places"
      "type": "object"
      "required":
        - "events"
        - "assets"
        - "places"
      "title": "DayContextSummaryCounts"
      "description": "Lightweight per-day counts for hover context panels."
...
    "ExplorationPerson":
      "properties":
        "id":
          "type": "integer"
          "title": "Id"
        "name":
          "type": "string"
          "title": "Name"
        "role":
          "anyOf":
            -
              "type": "string"
            -
              "type": "null"
          "title": "Role"
      "type": "object"
      "required":
        - "id"
        - "name"
        - "role"
      "title": "ExplorationPerson"
      "description": "Person item used by the exploration shell."
    "ExplorationRange":
      "properties":
        "start":
          "type": "string"
          "format": "date"
          "title": "Start"
        "end":
          "type": "string"
          "format": "date"
          "title": "End"
      "type": "object"
      "required":
        - "start"
        - "end"
      "title": "ExplorationRange"
      "description": "Resolved inclusive date range for exploration responses."
    "ExplorationTag":
      "properties":
        "path":
          "type": "string"
          "title": "Path"
        "label":
          "type": "string"
          "title": "Label"
      "type": "object"
      "required":
        - "path"
        - "label"
      "title": "ExplorationTag"
      "description": "Tag item used by the exploration shell."
...